backend/requirements.txt
asgiref==3.5.1
autopep8==1.6.0
beautifultable==0.8.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
distlib==0.3.4
Django==3.2.13
django-jalali-date==1.0.1
django-resized==1.0.0
djangorestframework==3.13.1
fandogh-cli==1.26.10
filelock==3.6.0
idna==3.3
jdatetime==4.1.0
pbr==5.8.1
Pillow==9.1.0
platformdirs==2.5.2
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==22.0.0
pytz==2022.1
PyYAML==5.4.1
requests==2.27.1
requests-toolbelt==0.9.1
retrying==1.3.3
six==1.16.0
sqlparse==0.4.2
stevedore==3.5.0
toml==0.10.2
tzlocal==2.1
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
websocket-client==1.3.2
